test_code/test_dq_check_short.py
```python
import pytest

from pyspark.sql import SparkSession
from pyspark.sql.types import StringType, IntegerType, DateType
from pyspark.sql.functions import col
from data_check import DataCheck

spark = SparkSession.builder.master("local").appName("DataCheck Test").getOrCreate()

data = [
    ("John Doe", "2021-12-01", 25, 100.5),
    ("Jane Smith", None, None, None),
    ("Alice", "2021-12-10", None, 80.0),
    ("Bob", "2021-11-30", 28, "invalid")
]

df = spark.createDataFrame(data, schema=("name", "date_of_birth", "age", "score"))

config_path = "path/to/config.json"
file_name = "sample_file"
src_system = "example"

@pytest.mark.parametrize("input_col,expected", [
    ("date_of_birth", ["None", "data_type_FAIL: Column [date_of_birth] should be DateType"]),
    ("age", ["None", "data_type_FAIL: Column [age] should be IntegerType"]),
])
def test_data_type_check(input_col, expected):
    data_check = DataCheck(source_df=df, spark_context=spark, config_path=config_path, file_name=file_name, src_system=src_system)
    data_check.data_type_check(input_col)
    result = data_check.source_df.select(f"`{input_col} type_check`").collect()
    for i in range(len(expected)):
        assert str(result[i][0]) == expected[i]

@pytest.mark.parametrize("input_col,nullable,expected", [
    ("name", True, ["None", "None", "None", "None"]),
    ("date_of_birth", False, ["None", "null_FAIL: Column [date_of_birth] cannot be null", "None", "None"]),
])
def test_null_check(input_col, nullable, expected):
    data_check = DataCheck(source_df=df, spark_context=spark, config_path=config_path, file_name=file_name, src_system=src_system)
    if not nullable:
        data_check.rule_df.at[input_col, "nullable"] = float("nan")
    data_check.null_check(input_col)
    result = data_check.source_df.select(f"`{input_col} null_check`").collect()
    for i in range(len(expected)):
        assert str(result[i][0]) == expected[i]

@pytest.mark.parametrize("input_col1,input_col2,syntax_value,expected", [
    ("age", "score", 125.5, True),
    ("age", "score", 150.0, False),
])
def test_sum_check_syntax(input_col1, input_col2, syntax_value, expected):
    data_check = DataCheck(source_df=df, spark_context=spark, config_path=config_path, file_name=file_name, src_system=src_system)
    sum_check_result = data_check.sum_check_syntax(input_col1, input_col2, syntax_value)
    result_rows = data_check.source_df.filter(sum_check_result).collect()
    assert bool(len(result_rows)) == expected
```

In the sample tests above:
- We use the `@pytest.mark.parametrize` decorator to run the same test function multiple times with different input and expected output values, making our tests more concise and readable.
- We use the PySpark DataFrame's `collect()` method to convert the result DataFrame into a list, which allows for easy comparison with the expected values.
- We use the `assert` statement to check if the actual and expected values are the same, and if not, the test will fail.

Running these unit tests with `pytest` will give clear information on which tests passed successfully and which ones failed, allowing you to easily identify issues in the application and refine the code as needed. A comprehensive test suite can help ensure the long-term stability and reliability of your application.